refactor(ui): migrate to Next.js App Router
